requirements.txt
requests==2.31.0
pandas==2.1.4
flask==3.0.0
flask-cors==4.0.0
redis==5.0.1
celery==5.3.4
waitress==3.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
cachetools==5.3.2
psutil==5.9.8 